// libs/content-ee/process-services-extension/src/lib/models/process-service.model.ts
/*
 * Copyright © 2005-2024 Hyland Software, Inc. and its affiliates. All rights reserved.
 *
 * License rights for this program may be obtained from Hyland Software, Inc.
 * pursuant to a written agreement and any use of this program without such an
 * agreement is prohibited.
 */

export interface UploadWidgetType {
    id: string;
    type: string;
    link: boolean;
}

export const ALL_APPS = 0;